requirements.txt
Django==4.1
djangorestframework==3.13.1
psycopg2>=2.7,<3.0